feat: skip vcs directory trees during search
These can have lots of files and should never produce a match, so not
bothering to iterate them can make things more efficient.

This also restructures the bounded concurrency model a bit to be more
simple, while extracting the initial filesystem walk code to make
it more modular and configurable in the future.